Hacked around a problem with glibc...
glibc broke some functionality (as in changed the default behaviors
without telling anyone outside of themselves on some of the placement
and use of some of their macros.  Cheap trick is to munge the macros
out of the first offending libc-symbols.h version that does the new
behavior so we can give people *SOMETHING* to work with while I figure
out just what in the *HELL* glibc's bunch did to me here and fix it
properly.