Initial steps for a "new" Lua version...
There's some "stupid" things going on in Scratchbox2 that makes it
at least ever so slightly more fragile than it ought to be.

1) We're cherry-picking Lua and building it right into libluaif
2) We're storing the source for THAT in our project.
3) We're using an OLD version of Lua (5.1.4...)- which could use
   being moved off to LuaJIT for some possible speed lifting or
   could be moved to 5.3 or later for speed, general cleanups
   (What's not to say the "memory leak" Nokia had observed
   in the comments in mailing lists for this wasn't due to the
   code in the old version?), and the like.

Can't do it without putzing around a bit.  So...we're plopping
LuaJIT 2.0.4 into here for now along with putting a line to
select between the two in the makefile as we sift through how
best to do this.  Might have it deleted, might not.  It's on
a branch to allow me to flat-out prune it out of existence.