First successful clean up item...
We removed liblua.a (Which, if you pause to think about it...you're
making a semi-custom framework here.  We also moved luaif to luafw
which is a more accurate way of describing this.  It's not done/over
yet on cleanups, restructures.  While one *CAN* make a claim for
Lua being in this project (so we can have a specifically built
static link engine that is built with -fPIC properly so it can live
in a .so as a piece part...), not all of this is clean, maintainable
as it currently exists.  So we'll work in stages on that next...