Close, close, close...
CMake build builds most of the stuff and packages/installs it right for
at least the host Arch level for binaries, etc.  We'll see, but this
is very, very close to being functional for the main host arch built
for this.  When it is done, we'll be able to likely use this to wrap
Yocto builds for things you JUST CAN'T PATCH OR TAPDANCE AROUND.